replace mongodb to postgresql

+++ src/main/kotlin/com/github/karolhor/ala/discounts/domain/model/Product.kt
@@ -1,10 +1,11 @@
 package com.github.karolhor.ala.discounts.domain.model
 
+import java.math.BigDecimal
 import java.util.UUID
 
 data class Product(
     val id: UUID,
     val name: String,
     val description: String,
-    val price: Int
+    val price: BigDecimal
 )
